observing password strength in home fragment

+++ app/src/main/java/com/robert/passwordmanager/models/Account.kt
@@ -13,7 +13,7 @@ data class Account(
     @ColumnInfo(name ="date")val date: String,
     @ColumnInfo(name ="user_name") val userName: String,
     @ColumnInfo(name ="website_name") val websiteName: String,
-    @ColumnInfo(name = "password_strength") val passwordStrength: Float
+    @ColumnInfo(name = "password_strength") val passwordStrength: Double
 ){
 
 }
